V021: 20-item review + hardening pass (v4.1.0)
Bugs fixed:
- Savings ledger section never rendered (_savings_tab referenced, not called)
- Row caps disabled by any SQL containing 'limit' (RATE_LIMIT columns);
  now word-boundary LIMIT detection in run()/run_batch()
- Python 'today' used server tz (UTC in SiS): MTD/forecast/contract-pace
  boundaries drifted a day every evening; account_today() (America/Chicago)
- Transient role-probe failure pinned session to ANALYST profile

Resilience/perf:
- 90s statement timeout on Cortex COMPLETE
- Usage logging + error sink write async (collect_nowait, sync fallback)
- Sidebar Refresh re-resolves role; health strip fetched/parsed once
- Altair theme via >=5.5 alt.theme API (deprecation warning gone)

Hardening/UX:
- exec_summary_html escapes every field; sidebar strip escapes text
- Expired-session errors map to a 'press Refresh data' hint
- safe_page captions name Python bug types (KeyError class of crash)
- Not-connected screen shows the underlying connection error
- .gitignore blocks .git_* recovery dirs

Tooling/docs:
- ruff widened: C4/SIM/PIE/PERF/RUF at zero findings
- CI: concurrency-cancel, pip cache, 15-min timeout
- CHANGELOG 4.1.0; APP_VERSION 4.1.0
- 17 regression tests (tests/test_hardening_v21.py); suite now 304

Includes the prior uncommitted V020 polish (FEATURES/RUNBOOK and page/data
touch-ups) that was already in the working tree.

+++ app/logic/governance.py
@@ -90,6 +90,6 @@ def governance_drift(inputs: dict, weights: dict | None = None) -> GovernanceDri
             "No auto-suspend", _cap(no_suspend * w["GOV_PTS_NO_AUTOSUSPEND"], 12),
             f"{no_suspend:.0f} warehouses without auto-suspend."))
 
-    score = int(round(max(0.0, 100.0 - sum(d.penalty for d in drivers))))
+    score = round(max(0.0, 100.0 - sum(d.penalty for d in drivers)))
     state = "Healthy" if score >= 90 else ("Watch" if score >= 75 else "Act")
     return GovernanceDrift(score=score, state=state, drivers=tuple(drivers))
